% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akrreport-package.R
\docType{package}
\name{akrreport-package}
\alias{akrreport-package}
\title{akrreport: Template for Alaska Refuge Reports}
\description{
Provides a template for using R Markdown to generate MS Word and PDF reports in the format of the Alaska Refuge Report Series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USFWS/akrreport}
  \item Report bugs at \url{https://github.com/USFWS/akrreport/issues}
}

}
\author{
\strong{Maintainer}: McCrea Cobb \email{mccrea_cobb@fws.gov}

}
\keyword{internal}
