% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quarto_doc.R
\name{create_quarto_doc}
\alias{create_quarto_doc}
\title{Create a new directory with the Quarto template}
\usage{
create_quarto_doc(dirname = "new_report", template = "word_doc")
}
\arguments{
\item{dirname}{character; the name of the directory to create.}

\item{template}{character; the template type to use. Currently includes a
template for MS Word documents.}
}
\description{
\code{create_quarto_doc} creates a new subdirectory inside the current directory, which will
contain the ready-to-use Quarto file and all associated files. The Word template
is based on the standard template of the Alaska Refuge Report Series.
}
\examples{
\dontrun{
 # Create template for Word document
 create_quarto_doc(dirname = "my_word_report", template = "word_doc")
}
}
