% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd_doc.R
\name{create_rmd_doc}
\alias{create_rmd_doc}
\title{Create a new directory with an R Markdown report template}
\usage{
create_rmd_doc(dirname = "new-doc", template = "word_doc")
}
\arguments{
\item{dirname}{Name of the directory to create.}

\item{template}{The name of the template to use.The current option is "word_doc".}
}
\description{
\code{create_rmd_doc} creates a new subdirectory inside the current directory,
which will contain the ready-to-use R Markdown file and associated files.
}
\details{
The function is a modified version of the \code{create.doc} function in the
\href{https://github.com/juba/rmdformats}{rmdformats} package.
}
\examples{
\dontrun{
# Create template for a MS Word report document
create_rmd_doc("my_report", template = "word_doc")
}
}
