% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_doc.R
\name{pdf_doc}
\alias{pdf_doc}
\title{Render a PDF report using R Markdown}
\usage{
pdf_doc(..., keep_tex = FALSE)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[rmarkdown]{pdf_document}}.}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{FALSE}}
}
\value{
output format to pass to \code{\link[bookdown:render_book]{render_book}}

output format to pass to \code{\link[bookdown:render_book]{render_book}}
}
\description{
This function serves as wrapper to \code{\link[rmarkdown]{pdf_document}}, with a
custom Pandoc \LaTeX template and different default values for other arguments
(e.g., \code{keep_tex = TRUE}) that match the template of the Alaska Refuge
Report Series.
}
\details{
Possible arguments for the YAML header are:
\itemize{
\item \code{bibliography} Path to the bibliography file to use for references
(BibTeX \emph{.bib} file). This template uses the bibliography-related
package \href{https://ctan.org/pkg/natbib}{natbib}. The current file
includes 3 dummy references; either insert your references into
this file or replace the file with your own.
\item \code{header-includes} Custom additions to the header, before the
\code{\\begin\{document\}} statement
\item \code{include-after} For including additional LaTeX code before the
\code{\\end\{document\}} statement
}
}
