% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_doc.R
\name{word_doc}
\alias{word_doc}
\title{Render a MS Word report using R Markdown}
\usage{
word_doc(
  toc = FALSE,
  toc_depth = 4,
  number_sections = FALSE,
  highlight = "default",
  reference_docx = "akrreport-template",
  dpi = 144,
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{toc}{logical; \code{TRUE} to include a table of contents in the output.}

\item{toc_depth}{integer; Depth of headers to include in table of contents. Default set to 4.}

\item{number_sections}{logical; whether to number section headers: if TRUE, figure/table numbers
will be of the form X.i, where X is the current first-level section number, and i is an
incremental number (the i-th figure/table); if FALSE (default), figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference section
headers in this case.}

\item{highlight}{character; syntax highlighting style. Supported styles include "default",
"tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock".
Pass \code{NULL} to prevent syntax highlighting.}

\item{reference_docx}{character; use the specified file as a style reference in producing a docx file.
The 'akrreport-template.docx' template implements the standard requirements for the Alaska
Refuge Report Series. If you prefer another template, pass the file name to this argument or simply use
'default' to use your standard Word template.}

\item{dpi}{integer; the resolution of the output figures, default is 144 dots per inch.}

\item{pandoc_args}{Additional command line options to pass to pandoc.}

\item{...}{Additional parameters to pass to \code{\link[bookdown]{word_document2}}.}
}
\value{
A modified \code{\link[rmarkdown]{word_document}} based on a Word template.
}
\description{
This function serves as wrapper for \code{\link[bookdown]{word_document2}},
with a custom Pandoc Word template and knitr default values
(e.g., \code{number_sections = FALSE}) that match the template of the Alaska
Refuge Report Series.
}
\examples{
\dontrun{
 # put in YAML header:
 output: akrreport::word_doc
}
}
